chore: update digger.yml (#25)
Upgrade to:
1. Take advantage of GH Issue workflow for drift mitigation. It ended up
being reverted so I didnt merge this 😔
See: https://diggertalk.slack.com/archives/C01G9JBJKQB/p1761055474978439
for context
2. Ensure module modifications trigger digger to replan (this didnt seem
to work because the path-includes argument seems to be ignored when
using terragrunt). <-- not sure if this only reads the config from `main`. Merging to test in a separate PR